fix(src): Risolto other methods
Riscritta la parte dei metodi Other Methods per adattarsi alla
nuova disposizione della libreria.

Generata documentazione
