--- cnlp/other_methods/information_theory/__init__.py
@@ -1,2 +1,0 @@
-from .neighbor_set_information import MI
-from .path_entropy import path_entropy
