+++ docs/other_methods/index.html
@@ -27,13 +27,13 @@
 <summary>
 <span>Expand source code</span>
 </summary>
-<pre><code class="python">from . import information_theory</code></pre>
+<pre><code class="python">import cnlp.other_methods.information_theory</code></pre>
 </details>
 </section>
 <section>
 <h2 class="section-title" id="header-submodules">Sub-modules</h2>
 <dl>
-<dt><code class="name"><a title="cnlp.other_methods.information_theory" href="information_theory/index.html">cnlp.other_methods.information_theory</a></code></dt>
+<dt><code class="name"><a title="cnlp.other_methods.information_theory" href="information_theory.html">cnlp.other_methods.information_theory</a></code></dt>
 <dd>
 <div class="desc"></div>
 </dd>
@@ -59,7 +59,7 @@
 </li>
 <li><h3><a href="#header-submodules">Sub-modules</a></h3>
 <ul>
-<li><code><a title="cnlp.other_methods.information_theory" href="information_theory/index.html">cnlp.other_methods.information_theory</a></code></li>
+<li><code><a title="cnlp.other_methods.information_theory" href="information_theory.html">cnlp.other_methods.information_theory</a></code></li>
 </ul>
 </li>
 </ul>
